docker-compose and refactor home screen
