+++ apps/expo/.env.development
@@ -1,0 +1,3 @@
+
+EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY=pk_test_ZXRoaWNhbC1zbmFwcGVyLTk0LmNsZXJrLmFjY291bnRzLmRldiQ
+EXPO_PUBLIC_API_URL="localhost:3000"
