+++ apps/expo/app.config.ts
@@ -32,6 +32,7 @@ const defineConfig = (): ExpoConfig => ({
     eas: {
       projectId: "e5b5d2cd-098b-4fe4-85ed-ac05e395552d", // dennis' project id for now
     },
+    clerkPublishableKey: process.env.CLERK_PUBLISHABLE_KEY,
   },
   experiments: {
     tsconfigPaths: true,
