+++ apps/expo/package.json
@@ -15,6 +15,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@clerk/clerk-expo": "^1.1.4",
     "@expo/metro-config": "^0.17.6",
     "@react-native-async-storage/async-storage": "1.21.0",
     "@react-native-community/datetimepicker": "7.6.1",
@@ -52,6 +53,7 @@
     "react-native-svg": "^15.1.0",
     "superjson": "2.2.1",
     "tamagui": "^1.94.4",
+    "zod": "^3.22.4",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
