+++ apps/expo/src/app/auth/index.tsx
@@ -1,0 +1,5 @@
+import { Text } from "tamagui";
+
+export default function Auth() {
+  return <Text>Auth</Text>;
+}
