+++ apps/expo/src/app/home/_components/station-tabs.tsx
@@ -1,0 +1,91 @@
+import { H3, ScrollView, Separator, Tabs, Text, YGroup, YStack } from "tamagui";
+
+import type { MenuWithRelations } from "@zotmeal/db";
+
+import { groupBy } from "~/utils";
+import { DishCard } from './dish-card';
+
+type Station = MenuWithRelations["stations"][0];
+type Dish = MenuWithRelations["stations"][0]["dishes"][0];
+
+export const StationTabs = ({
+  stations,
+}: Readonly<{ stations: Station[] }>) => (
+  <Tabs
+    defaultValue={stations?.[0]?.name}
+    orientation="horizontal"
+    flexDirection="column"
+    width={"100%"}
+    height={"100%"}
+  >
+    <Tabs.List>
+      <ScrollView
+        horizontal
+        bounces={false}
+        showsHorizontalScrollIndicator={false}
+      >
+        {stations.map((station) => (
+          <Tabs.Tab
+            key={station.name}
+            flex={1}
+            value={station.name}
+            borderRadius="$10"
+            height={"$3"}
+            marginHorizontal="$1"
+            marginBottom="$3"
+          >
+            <Text fontSize={"$5"}>{station.name}</Text>
+          </Tabs.Tab>
+        ))}
+      </ScrollView>
+    </Tabs.List>
+
+    {stations.map((station) => (
+      <Tabs.Content key={station.name} value={station.name}>
+        <ScrollView
+          padding="$2"
+          contentContainerStyle={{
+            alignItems: "center",
+          }}
+          contentInset={{
+            top: 0,
+            bottom: 200, // Make space at the bottom of the ScrollView
+          }}
+        >
+          {/* group dishes by category */}
+          {Object.entries(
+            groupBy(station.dishes, (dish) => dish.category as keyof Dish),
+          ).map(([category, dishes]) => (
+            <Category
+              key={category}
+              stationId={station.id}
+              category={category}
+              dishes={dishes}
+            />
+          ))}
+        </ScrollView>
+      </Tabs.Content>
+    ))}
+  </Tabs>
+);
+
+const Category = ({
+  stationId,
+  category,
+  dishes,
+}: Readonly<{
+  stationId: Station["id"];
+  category: string;
+  dishes: Dish[];
+}>) => (
+  <YStack key={category} width={"100%"}>
+    <H3 fontWeight={"800"} marginTop="$5" paddingLeft="$2">
+      {category}
+    </H3>
+    <YGroup bordered separator={<Separator borderWidth={1} />}>
+      {dishes.map((dish) => (
+        <DishCard key={dish.id} dish={dish} stationId={stationId} />
+      ))}
+    </YGroup>
+  </YStack>
+);
