+++ packages/api/src/schedules/getWeekInfo.ts
