--- packages/api/src/services/index.ts
@@ -1,2 +1,0 @@
-export * from "./getWeekInfo";
-export * from "./updateDaily";
